feat: introduce API Gateway authorizer function BREAKING CHANGE: renamed Handler and related types to FunctionHandler
